Move continuous build classes in launcher to package
