+++ subprojects/launcher/src/main/java/org/gradle/tooling/internal/provider/continuous/AccumulateBuildInputsListener.java
@@ -14,14 +14,13 @@
  * limitations under the License.
  */
 
-package org.gradle.tooling.internal.provider;
+package org.gradle.tooling.internal.provider.continuous;
 
 import org.gradle.api.internal.file.FileCollectionInternal;
 import org.gradle.api.internal.file.FileCollectionStructureVisitor;
 import org.gradle.api.internal.file.FileTreeInternal;
 import org.gradle.api.internal.file.collections.FileSystemMirroringFileTree;
 import org.gradle.api.tasks.util.PatternSet;
-import org.gradle.execution.plan.BuildInputHierarchy;
 import org.gradle.internal.execution.UnitOfWork;
 import org.gradle.internal.execution.WorkInputListener;
 import org.gradle.internal.execution.fingerprint.InputFingerprinter;
